fix: 修正Display Entity的Billboard数据类型为byte
修正实体元数据中Billboard渲染约束的数据类型，从int改为byte以匹配NMS实现

+++ src/main/java/com/oolonghoo/holograms/nms/versions/EntityMetadataBuilder.java
@@ -259,7 +259,7 @@ public class EntityMetadataBuilder {
         if (billboard == null) {
             billboard = Billboard.CENTER;
         }
-        int billboardValue;
+        byte billboardValue;
         switch (billboard) {
             case FIXED:
                 billboardValue = 0;
